requirements.txt
easydict==1.13
numpy==1.21.6
pandas==1.3.5
Pillow==9.5.0
Pillow==10.4.0
PyYAML==6.0.1
PyYAML==6.0.1
torch==1.10.0+cu111
torchvision==0.11.0+cu111
